Ajustes crud dibujos

+++ src/main/java/com/example/proyecto/domain/Dibujo.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @EqualsAndHashCode(of="id")
 public class Dibujo {
 
-    private Long id;
+    private BigDecimal id;
 
     @NotNull
     private String titulo;
